+++ skrub/_dataframe/_test_utils.py
@@ -1,0 +1,6 @@
+def is_namespace_pandas(px):
+    return px.__name__ == "pandas"
+
+
+def is_namespace_polars(px):
+    return px.__name__ == "polars"
