+++ skrub/tests/test_interpolation_join.py
@@ -7,7 +7,7 @@ from sklearn.neighbors import KNeighborsClassifier, KNeighborsRegressor
 
 from skrub import InterpolationJoiner
 from skrub._dataframe._polars import POLARS_SETUP
-from skrub._utils import is_namespace_polars
+from skrub._dataframe._test_utils import is_namespace_polars
 
 MODULES = [pd]
