+++ skrub/tests/test_joiner.py
@@ -5,7 +5,7 @@ from pandas.testing import assert_frame_equal
 
 from skrub import Joiner
 from skrub._dataframe._polars import POLARS_SETUP
-from skrub._utils import is_namespace_polars
+from skrub._dataframe._test_utils import is_namespace_polars
 
 MODULES = [pd]
 ASSERT_TUPLES = [(pd, assert_frame_equal)]
