+++ skrub/tests/test_similarity_encoder.py
@@ -8,9 +8,9 @@ from sklearn.exceptions import NotFittedError
 
 from skrub import SimilarityEncoder
 from skrub._dataframe._polars import POLARS_SETUP
+from skrub._dataframe._test_utils import is_namespace_polars
 from skrub._similarity_encoder import ngram_similarity_matrix
 from skrub._string_distances import ngram_similarity
-from skrub._utils import is_namespace_polars
 
 MODULES = [pd]
 INPUT_TYPES = ["list", "numpy", "pandas"]
